pyproject.toml
[project]
name = "3dscenesfm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.1"
dependencies = [
    "dvc[gdrive]>=3.60.0",
    "kornia>=0.8.1",
    "lightglue",
    "lightning>=2.5.1.post0",
    "numpy==1.26.4",
    "open3d>=0.19.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pycolmap>=3.11.1",
    "torch>=2.2.2",
    "transformers>=4.52.4",
]

[tool.uv.sources]
lightglue = { git = "https://github.com/cvg/LightGlue.git" }

[dependency-groups]
dev = [
    "fastprogress>=1.0.3",
    "fire>=0.7.0",
    "h5py>=3.13.0",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "mlflow>=3.1.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.13",
    "tqdm>=4.67.1",
]
